Split vultron/wire/as2/extractor.py into focused submodules
Converts the 826-line extractor.py into a proper package with four
focused submodules, all under the 500-line limit (CS-18-002):

- extractor/_pattern.py: ActivityPattern class + _match_activity_field
- extractor/_instances.py: all *Pattern module-level constants
- extractor/_builders.py: field-extraction helpers + domain-object
  builders + _build_object_kwargs dispatch
- extractor/_extract.py: extract_intent function

The package __init__.py re-exports all 46 public names (ActivityPattern,
extract_intent, and all *Pattern instances) so all existing callers in
vultron/semantic_registry/ and tests continue to work without change
(AC-4, CS-18-003).

The _validate_registry_order() import-time order guard remains active
via vultron/semantic_registry/__init__.py (AC-5).

Also updates vultron/wire/as2/AGENTS.md file-locations table and the
adding-a-pattern checklist to reflect the new package layout.

Closes #880
