+++ vultron/wire/as2/extractor/_builders.py
@@ -1,358 +1,32 @@
-"""AS2 wire layer semantic extractor for the Vultron Protocol.
+"""Private domain-object builder helpers for the Vultron extractor.
 
-Maps AS2 activity structures to domain MessageSemantics. This is the sole
-location where AS2 vocabulary is translated to domain concepts (ARCH-03-001).
-This is stage 3 of the inbound pipeline: typed AS2 activity → MessageSemantics.
-
-Consolidates ActivityPattern definitions (formerly in vultron/activity_patterns.py)
-into a single module.  Pattern-to-semantics matching lives in
-``vultron.semantic_registry``, which iterates ``SEMANTIC_REGISTRY`` directly.
+Converts raw AS2 wire objects to core domain model instances.  All
+functions here are internal implementation details of
+``vultron.wire.as2.extractor``; callers should use ``extract_intent``
+from ``_extract.py`` (or the ``vultron.semantic_registry.extract_event``
+convenience wrapper) rather than calling these builders directly.
 """
 
 from datetime import datetime
-from typing import Any, Callable, Optional, Union
+from typing import Any, Callable
 
-from pydantic import BaseModel
-
-from vultron.wire.as2.vocab.base.objects.activities.base import as_Activity
-from vultron.wire.as2.vocab.base.objects.actors import as_Actor
-from vultron.wire.as2.vocab.base.objects.object_types import as_Event
-from vultron.core.models.actor import CoreActor
-from vultron.core.models.base import VultronObject
 from vultron.core.models._helpers import _now_utc as _core_now_utc
+from vultron.core.models.base import VultronObject
 from vultron.core.models.case_log_entry import VultronCaseLogEntry
-from vultron.core.models.events import (
-    MessageSemantics,
-    VultronEvent,
-)
 from vultron.core.models.enums import VultronObjectType as VOtype
 from vultron.core.models.vultron_types import (
-    VultronActivity,
-    VultronCase,
     CaseStatus,
     EmbargoEvent,
+    ParticipantStatus,
+    VultronActivity,
+    VultronCase,
     VultronNote,
     VultronParticipant,
-    ParticipantStatus,
     VultronReport,
 )
-from vultron.wire.as2.enums import (
-    as_IntransitiveActivityType as IAtype,
-    as_ObjectType as AOtype,
-    as_TransitiveActivityType as TAtype,
-)
-
-
-class ActivityPattern(BaseModel):
-    """Represents a pattern to match against an AS2 activity for semantic dispatch.
-
-    Supports nested patterns for activities whose object is itself an activity.
-    """
-
-    description: Optional[str] = None
-    activity_: TAtype | IAtype
-
-    to_: Optional[Union[AOtype, VOtype, "ActivityPattern"]] = None
-    object_: Optional[Union[AOtype, VOtype, "ActivityPattern"]] = None
-    target_: Optional[Union[AOtype, VOtype, "ActivityPattern"]] = None
-    context_: Optional[Union[AOtype, VOtype, "ActivityPattern"]] = None
-    in_reply_to_: Optional["ActivityPattern"] = None
-
-    def match(self, activity: as_Activity) -> bool:
-        """Return True if the given activity matches this pattern."""
-        if self.activity_ != activity.type_:
-            return False
-
-        field_pairs = (
-            (self.object_, getattr(activity, "object_", None)),
-            (self.target_, getattr(activity, "target", None)),
-            (self.context_, getattr(activity, "context", None)),
-            (self.to_, getattr(activity, "to", None)),
-            (self.in_reply_to_, getattr(activity, "in_reply_to", None)),
-        )
-        return all(
-            _match_activity_field(pattern_field, activity_field)
-            for pattern_field, activity_field in field_pairs
-        )
-
-
-def _match_activity_field(
-    pattern_field: AOtype | VOtype | ActivityPattern | None,
-    activity_field: object,
-) -> bool:
-    if pattern_field is None:
-        return True
-    if isinstance(pattern_field, ActivityPattern):
-        return isinstance(activity_field, as_Activity) and pattern_field.match(
-            activity_field
-        )
-    if isinstance(activity_field, str):
-        return True
-    if activity_field is None:
-        return False
-    if pattern_field == AOtype.ACTOR and isinstance(
-        activity_field, (as_Actor, CoreActor)
-    ):
-        return True
-    if pattern_field == AOtype.EVENT and isinstance(activity_field, as_Event):
-        return True
-    return bool(pattern_field == getattr(activity_field, "type_", None))
-
-
-# ---------------------------------------------------------------------------
-# Pattern instances (formerly in vultron/activity_patterns.py)
-# ---------------------------------------------------------------------------
-
-CreateEmbargoEventPattern = ActivityPattern(
-    description=(
-        "Create an embargo event. This is the initial step in the embargo "
-        "management process, where a coordinator creates an embargo event to "
-        "manage the embargo on a vulnerability case."
-    ),
-    activity_=TAtype.CREATE,
-    object_=AOtype.EVENT,
-    context_=VOtype.VULNERABILITY_CASE,
-)
-AddEmbargoEventToCasePattern = ActivityPattern(
-    description=(
-        "Add an embargo event to a vulnerability case. This is typically "
-        "observed as an ADD activity where the object is an EVENT and the "
-        "target is a VULNERABILITY_CASE."
-    ),
-    activity_=TAtype.ADD,
-    object_=AOtype.EVENT,
-    target_=VOtype.VULNERABILITY_CASE,
-)
-RemoveEmbargoEventFromCasePattern = ActivityPattern(
-    description=(
-        "Remove an embargo event from a vulnerability case. This is typically "
-        "observed as a REMOVE activity where the object is an EVENT. The "
-        "origin field of the activity contains the VulnerabilityCase from "
-        "which the embargo is removed."
-    ),
-    activity_=TAtype.REMOVE,
-    object_=AOtype.EVENT,
-)
-AnnounceEmbargoEventToCasePattern = ActivityPattern(
-    description=(
-        "Announce an embargo event to a vulnerability case. This is typically "
-        "observed as an ANNOUNCE activity where the object is an EVENT and the "
-        "context is a VULNERABILITY_CASE."
-    ),
-    activity_=TAtype.ANNOUNCE,
-    object_=AOtype.EVENT,
-    context_=VOtype.VULNERABILITY_CASE,
-)
-InviteToEmbargoOnCasePattern = ActivityPattern(
-    description=(
-        "Propose an embargo on a vulnerability case. "
-        "This is observed as an INVITE activity where the object is an "
-        "EmbargoEvent and the context is the VulnerabilityCase. "
-        "Corresponds to EmProposeEmbargoActivity."
-    ),
-    activity_=TAtype.INVITE,
-    object_=AOtype.EVENT,
-    context_=VOtype.VULNERABILITY_CASE,
-)
-AcceptInviteToEmbargoOnCasePattern = ActivityPattern(
-    description="Accept an invitation to an embargo on a vulnerability case.",
-    activity_=TAtype.ACCEPT,
-    object_=InviteToEmbargoOnCasePattern,
-)
-RejectInviteToEmbargoOnCasePattern = ActivityPattern(
-    description="Reject an invitation to an embargo on a vulnerability case.",
-    activity_=TAtype.REJECT,
-    object_=InviteToEmbargoOnCasePattern,
-)
-CreateReportPattern = ActivityPattern(
-    description=(
-        "Create a vulnerability report. This is the initial step in the "
-        "vulnerability disclosure process, where a finder creates a report to "
-        "disclose a vulnerability. It may not always be observed directly, as "
-        "it could be implicit in the OFFER of the report."
-    ),
-    activity_=TAtype.CREATE,
-    object_=VOtype.VULNERABILITY_REPORT,
-)
-ReportSubmissionPattern = ActivityPattern(
-    description=(
-        "Submit a vulnerability report for validation. This is typically "
-        "observed as an OFFER of a VULNERABILITY_REPORT, which represents the "
-        "submission of the report to a coordinator or vendor for validation."
-    ),
-    activity_=TAtype.OFFER,
-    object_=VOtype.VULNERABILITY_REPORT,
-)
-AckReportPattern = ActivityPattern(
-    activity_=TAtype.READ, object_=ReportSubmissionPattern
-)
-ValidateReportPattern = ActivityPattern(
-    activity_=TAtype.ACCEPT, object_=ReportSubmissionPattern
-)
-InvalidateReportPattern = ActivityPattern(
-    activity_=TAtype.TENTATIVE_REJECT, object_=ReportSubmissionPattern
-)
-CloseReportPattern = ActivityPattern(
-    activity_=TAtype.REJECT, object_=ReportSubmissionPattern
-)
-CreateCaseActivityPattern = ActivityPattern(
-    activity_=TAtype.CREATE, object_=VOtype.VULNERABILITY_CASE
-)
-UpdateCaseActivityPattern = ActivityPattern(
-    activity_=TAtype.UPDATE, object_=VOtype.VULNERABILITY_CASE
-)
-EngageCasePattern = ActivityPattern(
-    description=(
-        "Actor engages (joins) a VulnerabilityCase, transitioning their RM "
-        "state to ACCEPTED."
-    ),
-    activity_=TAtype.JOIN,
-    object_=VOtype.VULNERABILITY_CASE,
-)
-DeferCasePattern = ActivityPattern(
-    description=(
-        "Actor defers (ignores) a VulnerabilityCase, transitioning their RM "
-        "state to DEFERRED."
-    ),
-    activity_=TAtype.IGNORE,
-    object_=VOtype.VULNERABILITY_CASE,
-)
-AddReportToCaseActivityPattern = ActivityPattern(
-    activity_=TAtype.ADD,
-    object_=VOtype.VULNERABILITY_REPORT,
-    target_=VOtype.VULNERABILITY_CASE,
-)
-SuggestActorToCasePattern = ActivityPattern(
-    activity_=TAtype.OFFER,
-    object_=AOtype.ACTOR,
-    target_=VOtype.VULNERABILITY_CASE,
-)
-AcceptSuggestActorToCasePattern = ActivityPattern(
-    activity_=TAtype.ACCEPT, object_=SuggestActorToCasePattern
-)
-RejectSuggestActorToCasePattern = ActivityPattern(
-    activity_=TAtype.REJECT, object_=SuggestActorToCasePattern
-)
-OfferCaseManagerRolePattern = ActivityPattern(
-    description=(
-        "Vendor offers the CASE_MANAGER role to a Case Actor participant. "
-        "Distinct from OFFER_CASE_OWNERSHIP_TRANSFER: the offering actor "
-        "retains CASE_OWNER; only operational management authority is "
-        "delegated. Identified by target being a CASE_PARTICIPANT record. "
-        "See DEMOMA-08-002, DEMOMA-08-003."
-    ),
-    activity_=TAtype.OFFER,
-    object_=VOtype.VULNERABILITY_CASE,
-    target_=VOtype.CASE_PARTICIPANT,
-)
-AcceptCaseManagerRolePattern = ActivityPattern(
-    description="Case Actor accepted the CASE_MANAGER role delegation offer.",
-    activity_=TAtype.ACCEPT,
-    object_=OfferCaseManagerRolePattern,
-)
-RejectCaseManagerRolePattern = ActivityPattern(
-    description="Case Actor rejected the CASE_MANAGER role delegation offer.",
-    activity_=TAtype.REJECT,
-    object_=OfferCaseManagerRolePattern,
-)
-OfferCaseOwnershipTransferActivityPattern = ActivityPattern(
-    activity_=TAtype.OFFER, object_=VOtype.VULNERABILITY_CASE
-)
-AcceptCaseOwnershipTransferActivityPattern = ActivityPattern(
-    activity_=TAtype.ACCEPT, object_=OfferCaseOwnershipTransferActivityPattern
-)
-RejectCaseOwnershipTransferActivityPattern = ActivityPattern(
-    activity_=TAtype.REJECT, object_=OfferCaseOwnershipTransferActivityPattern
-)
-InviteActorToCasePattern = ActivityPattern(
-    activity_=TAtype.INVITE,
-    object_=AOtype.ACTOR,
-    target_=VOtype.VULNERABILITY_CASE,
-)
-AcceptInviteActorToCasePattern = ActivityPattern(
-    activity_=TAtype.ACCEPT,
-    object_=InviteActorToCasePattern,
-)
-RejectInviteActorToCasePattern = ActivityPattern(
-    activity_=TAtype.REJECT,
-    object_=InviteActorToCasePattern,
-)
-CloseCasePattern = ActivityPattern(
-    activity_=TAtype.LEAVE, object_=VOtype.VULNERABILITY_CASE
-)
-AnnounceLogEntryPattern = ActivityPattern(
-    description=(
-        "Announce a canonical CaseLogEntry to a participant for log "
-        "replication. The object is a CaseLogEntry object."
-    ),
-    activity_=TAtype.ANNOUNCE,
-    object_=VOtype.CASE_LOG_ENTRY,
-)
-AnnounceVulnerabilityCasePattern = ActivityPattern(
-    description=(
-        "Case owner announces full VulnerabilityCase details to a newly "
-        "accepted invitee (MV-10-003). The object is a VulnerabilityCase."
-    ),
-    activity_=TAtype.ANNOUNCE,
-    object_=VOtype.VULNERABILITY_CASE,
-)
-RejectLogEntryPattern = ActivityPattern(
-    description=(
-        "Participant rejects a CaseLogEntry announcement due to "
-        "hash-chain mismatch. The object is the rejected CaseLogEntry."
-    ),
-    activity_=TAtype.REJECT,
-    object_=VOtype.CASE_LOG_ENTRY,
-)
-CreateNotePattern = ActivityPattern(
-    activity_=TAtype.CREATE,
-    object_=AOtype.NOTE,
-)
-AddNoteToCaseActivityPattern = ActivityPattern(
-    activity_=TAtype.ADD,
-    object_=AOtype.NOTE,
-    target_=VOtype.VULNERABILITY_CASE,
-)
-RemoveNoteFromCasePattern = ActivityPattern(
-    activity_=TAtype.REMOVE,
-    object_=AOtype.NOTE,
-    target_=VOtype.VULNERABILITY_CASE,
-)
-CreateCaseParticipantPattern = ActivityPattern(
-    activity_=TAtype.CREATE,
-    object_=VOtype.CASE_PARTICIPANT,
-    context_=VOtype.VULNERABILITY_CASE,
-)
-AddCaseParticipantToCasePattern = ActivityPattern(
-    activity_=TAtype.ADD,
-    object_=VOtype.CASE_PARTICIPANT,
-    target_=VOtype.VULNERABILITY_CASE,
-)
-RemoveCaseParticipantFromCasePattern = ActivityPattern(
-    activity_=TAtype.REMOVE,
-    object_=VOtype.CASE_PARTICIPANT,
-    target_=VOtype.VULNERABILITY_CASE,
-)
-CreateCaseStatusActivityPattern = ActivityPattern(
-    activity_=TAtype.CREATE,
-    object_=VOtype.CASE_STATUS,
-    context_=VOtype.VULNERABILITY_CASE,
-)
-AddCaseStatusToCasePattern = ActivityPattern(
-    activity_=TAtype.ADD,
-    object_=VOtype.CASE_STATUS,
-    target_=VOtype.VULNERABILITY_CASE,
-)
-CreateParticipantStatusPattern = ActivityPattern(
-    activity_=TAtype.CREATE,
-    object_=VOtype.PARTICIPANT_STATUS,
-)
-AddParticipantStatusToParticipantPattern = ActivityPattern(
-    activity_=TAtype.ADD,
-    object_=VOtype.PARTICIPANT_STATUS,
-    target_=VOtype.CASE_PARTICIPANT,
-)
-
+from vultron.wire.as2.enums import as_ObjectType as AOtype
+from vultron.wire.as2.vocab.base.objects.activities.base import as_Activity
+from vultron.wire.as2.vocab.base.objects.object_types import as_Event
 
 # ---------------------------------------------------------------------------
 # Private field-extraction helpers
@@ -755,72 +429,3 @@ def _build_object_kwargs(
         if obj_id:
             kw["object_"] = VultronObject(id_=obj_id, type_=_get_type(obj))
     return kw
-
-
-def extract_intent(
-    activity: as_Activity,
-    semantics: MessageSemantics,
-    event_class: type[VultronEvent],
-    include_activity: bool = False,
-) -> VultronEvent:
-    """Extract domain fields from an AS2 activity given pre-computed semantics.
-
-    This function is the sole AS2 → domain translation point.  It is called
-    after pattern matching has already determined the ``semantics``; the
-    caller must supply the matching ``event_class`` and ``include_activity``
-    flag from the registry entry.
-
-    For a single-call convenience wrapper that performs pattern matching and
-    registry lookup automatically, use ``vultron.semantic_registry.extract_event``.
-
-    Args:
-        activity: The AS2 activity to extract fields from.
-        semantics: Pre-matched ``MessageSemantics`` value.
-        event_class: Concrete ``VultronEvent`` subclass to instantiate.
-        include_activity: When ``True``, populate ``event.activity`` with a
-            summarised ``VultronActivity`` snapshot of the outer activity.
-
-    Returns:
-        A concrete VultronEvent subclass discriminated by MessageSemantics.
-    """
-    actor_id = _get_id(getattr(activity, "actor", None)) or ""
-    obj = getattr(activity, "object_", None)
-    target = getattr(activity, "target", None)
-    context = getattr(activity, "context", None)
-    origin = getattr(activity, "origin", None)
-
-    # Nested fields from activity.object_ (for Accept/Reject wrapping another activity)
-    inner_obj = inner_target = inner_context = None
-    if obj is not None and not isinstance(obj, str):
-        inner_obj = getattr(obj, "object_", None)
-        inner_target = getattr(obj, "target", None)
-        inner_context = getattr(obj, "context", None)
-
-    _obj_type = str(getattr(obj, "type_", "")) if obj is not None else ""
-    extra_kwargs = _build_object_kwargs(
-        obj,
-        _obj_type,
-        context,
-        target,
-        include_activity,
-        activity,
-        actor_id,
-        origin,
-    )
-    return event_class(
-        semantic_type=semantics,
-        activity_id=activity.id_,
-        activity_type=str(activity.type_) if activity.type_ else None,
-        actor_id=actor_id,
-        # object_ comes from extra_kwargs if a typed domain object was built;
-        # otherwise fall back to a minimal VultronObject wrapper.
-        object_=extra_kwargs.pop("object_", None) or _to_domain_obj(obj),
-        target=_to_domain_obj(target),
-        context=_to_domain_obj(context),
-        origin=_to_domain_obj(origin),
-        inner_object=_to_domain_obj(inner_obj),
-        inner_target=_to_domain_obj(inner_target),
-        inner_context=_to_domain_obj(inner_context),
-        in_reply_to=_get_id(getattr(activity, "in_reply_to", None)),
-        **extra_kwargs,
-    )
